Move WOWS distance finish line
