+++ src/wideopendistanceform.h
@@ -28,9 +28,6 @@ private:
     Ui::WideOpenDistanceForm *ui;
     MainWindow  *mMainWindow;
 
-    static void intercept(double lata1, double lona1, double lata2, double lona2,
-                          double latb1, double lonb1, double &lat0, double &lon0);
-
 public slots:
     void updateView();
